+++ internal/component/discovery/snmp/snmp.go
@@ -14,8 +14,9 @@ import (
 	"github.com/grafana/alloy/internal/component/discovery"
 	"github.com/grafana/alloy/internal/featuregate"
 	"github.com/grafana/alloy/internal/service/livedebugging"
-	"github.com/grafana/alloy/internal/snmpdiscovery"
 	"github.com/grafana/alloy/internal/snmppaths"
+
+	"github.com/Mesverrum/snmp-sd/snmpdiscovery"
 	"github.com/grafana/alloy/syntax/alloytypes"
 )